requirements.txt
requests~=2.27.1
colorama~=0.4.4
SQLAlchemy~=1.4.26
python-dotenv~=0.19.2